refactor: centralize shared types and simplify logger
